requirements.txt
python-telegram-bot==20.7
replicate==0.25.1
python-dotenv==1.0.1
httpx==0.27.0
urllib3==2.2.1